requirements-tests.txt
python-multipart
requests
pytest
bcrypt
psycopg2-binary
fastapi
